w2-analyzer-project/requirements.txt
# =============================================
# W-2 ANALYZER PROJECT - COMPLETE DEPENDENCIES
# =============================================

# CORE PYTHON & WEB FRAMEWORK
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0

# AI & MACHINE LEARNING
google-generativeai==0.3.2
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
pillow==10.1.0
pytesseract==0.3.10
pypdf2==3.0.1
pdf2image==1.16.3

# TESTING & DEVELOPMENT
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.27.0

# OPTIONAL OCR DEPENDENCIES (for better PDF/image processing)
# Install system Tesseract OCR for your platform:

# Windows: 
# Download from: https://github.com/UB-Mannheim/tesseract/wiki
# Add to PATH: C:\Program Files\Tesseract-OCR

# macOS:
# brew install tesseract

# Linux (Ubuntu/Debian):
# sudo apt-get install tesseract-ocr

# =============================================
# VERSION NOTES
# =============================================
# Python 3.9+ required
# Some packages may have compatibility notes:
# - google-generativeai: May show Python 3.9 EOL warning (works fine)
# - grpcio: Required for Google AI services
# - pytesseract: Requires system Tesseract installation